fixed bug 9.5

+++ src/utils/ajax/ajax.js
@@ -97,7 +97,7 @@ export default function({
 		}).catch((err) => {
 			console.dir(err);
 			this.$store.dispatch('hide_loading');
-			if (err.response.data === 'Error: Not allowed') {
+			if (err.response && err.response.data === 'Error: Not allowed') {
 				this.$store.dispatch('remove_userinfo').then(() => {
 					this.$router.push('/login');
 				});
